third_party/blink/web_tests/fast/forms/radio/state-restore-radio-group-expected.txt

Test to restore form value states for a radio group.

PASS document.getElementById("input1").checked is true
PASS document.getElementById("input2").checked is false
PASS successfullyParsed is true

TEST COMPLETE

 
